A classe Nota foi movida para dentro da classe Disciplina
