+++ src/main/kotlin/Disciplina.kt
@@ -1,9 +1,28 @@
 package org.example
 
+class Disciplina(
+    val nome: String,
+    val professor: String
+) {
 
-class Disciplina (
+    class Nota(
+        val somativa1: Double,
+        val somativa2: Double,
+        val avaliacao: Double,
+        val substitutiva: Double
+    ) {
 
-    val nome : String,
-    val professor : String,
+        fun getMedia(): Double {
+            var prova = avaliacao
 
-)
+            if (prova < substitutiva) {
+                prova = substitutiva
+            }
+
+            val media = (somativa1 * 0.2) + (somativa2 * 0.4) + (prova * 0.4)
+
+            return Math.round(media * 10) / 10.0
+        }
+    }
+}
+
